pyproject.toml
[build-system]
requires = [
    "scikit-build-core>=0.3.3",
    "pybind11>=2.6.0",
]
build-backend = "scikit_build_core.build"

[project]
name = "blackmagic-io"
version = "0.16.0b0"
description = "Python library for video I/O with Blackmagic DeckLink devices"
readme = "README.md"
requires-python = ">=3.7"
authors = [
    {name = "Nick Shaw", email = "nick@antlerpost.com"}
]
maintainers = [
    {name = "Nick Shaw", email = "nick@antlerpost.com"}
]
keywords = ["blackmagic", "decklink", "video", "output", "input", "io", "sdi", "hdmi", "broadcast"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Multimedia :: Video",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: C++",
    "Operating System :: OS Independent",
]
dependencies = [
    "numpy>=1.19.0,<2.3",
    "pybind11>=2.6.0",
]

[project.urls]
Homepage = "https://github.com/nick-shaw/blackmagic-output"
Repository = "https://github.com/nick-shaw/blackmagic-output"
Issues = "https://github.com/nick-shaw/blackmagic-output/issues"

[tool.scikit-build]
# Minimum CMake version
minimum-version = "0.3.3"
# Build directory for cmake artifacts
build-dir = "build/{wheel_tag}"
# Include vendor directory in sdist
sdist.include = ["_vendor/**"]

[dependency-groups]
dev = [
    "black>=23.3.0",
    "cibuildwheel>=2.13.1",
    "mypy>=1.4.1",
    "pytest>=7.4.4",
]
examples = [
    "imageio>=2.0.0",
    "pillow>=8.0.0",
]
# Verbose output during build
